+++ CHANGELOG.md
@@ -11,6 +11,7 @@ modified Semantic Versioning scheme. See the "Versioning scheme" section of the
 ## [Unreleased]
 ### Added
 - Added the [ACSF module](https://www.drupal.org/project/acsf) as a required dependency.
+- Added the [openid_connect](https://www.drupal.org/project/openid_connect) module as a dependency.
 
 ### Changed